revert(cosmos3): UniPC lambda back to float32 to match the reference
e5483de raised lambda = log(1-sigma) - log(sigma) to double on the theory
that float32 log(1-sigma) at sigma ~ 0.9998 was poisoning the order-2
divided difference. The reference computes it in float32 as well --
fm_solvers_unipc.py multistep_uni_p_bh_update calls torch.log on
self.sigmas, which set_timesteps stores as float32 -- so the precision
loss is part of the numerics we are trying to reproduce, and computing in
double makes this solver disagree with the reference instead of agreeing.

The dim0 / dim5 proprioception sweeps on the double build are unchanged
(0.589 / 0.399 vs 0.587 / 0.397 before it), confirming this was never the
gain deficit. The JSON-duration half of e5483de is kept: it aligns the
prompt duration with the reference video.shape[1] and renders identically
today.

+++ experimental_models/cosmos3/cpp/action/cosmos3Scheduler.cpp
@@ -37,12 +37,14 @@ float applyShift(double sigma, double shift)
 }
 
 //! Flow-matching half-log-SNR: lambda = log(alpha) - log(sigma), alpha = 1 - sigma.
-//! Computed in double: at the first UniPC step sigma ~ 0.9998, float32 log(1-sigma)
-//! has only a few bits and poisons the order-2 divided difference.
+//! Kept in float32 to match the reference, which evaluates torch.log on its float32
+//! sigma buffer. At the first UniPC step sigma ~ 0.9998 leaves log(1-sigma) with only
+//! a few significant bits, but that precision loss is part of the reference numerics:
+//! computing it in double here would make this solver disagree with the reference.
 double lambdaOf(double sigma)
 {
-    double const alpha = 1.0 - sigma;
-    return std::log(alpha) - std::log(sigma);
+    float const alpha = 1.0F - static_cast<float>(sigma);
+    return static_cast<double>(std::log(alpha) - std::log(static_cast<float>(sigma)));
 }
 } // namespace
